01.01.2020 20.30 We have connection with data base!

+++ src/main/java/HelloService.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public class HelloService {
     static final String FALLBACK_NAME="world";
-    static final Lang FALLBACK_LANG= new Lang(1L,"Hello","en");
+    static final Lang FALLBACK_LANG= new Lang(1,"Hello","en");
     private final Logger logger = LoggerFactory.getLogger(HelloService.class);
     private LangRepository repository;
     HelloService()
@@ -22,9 +22,9 @@ public class HelloService {
 
     String prepareGreeting(String name, String lang)
     {
-        Long langId;
+        Integer langId;
         try{
-            langId = Optional.ofNullable(lang).map(Long::valueOf).orElse(FALLBACK_LANG.getId());
+            langId = Optional.ofNullable(lang).map(Integer::valueOf).orElse(FALLBACK_LANG.getId());
         }
         catch (NumberFormatException e)
         {
